chore: move errors to a new package
